Simplify code and reduce gas costs.
When setting >250 roles, gas cost is reduced from ~12.2m to ~11.7m.

This is less of a reduction than I was hoping for, but it turns out that just tracking the user counts per role is a fundamental change that almost doubles our gas cost (which makes sense as it almost doubles our writes) from ~6.5m to ~11.7m.

For that reason I'm keeping these last two commits separate while we discuss whether or not we really want them.